Singularity has multiple meanings, including a mathematical concept, a theoretical location in space-time, and a video game:
Mathematics
A point where a mathematical object is not defined, or where it stops behaving well in some way. For example, the value of an equation that uses 1/X goes to infinity as X goes to zero.

Physics
A location in space-time where the laws of physics break down, such as when the gravitational field and density of a celestial body become infinite. This is also known as a gravitational or space-time singularity.

Video game
A multiplayer game where players fight as soldiers or creatures. In one mode, soldiers try to take over beacons on a map, while creatures defend them. When time runs out or the soldiers capture all the beacons, players switch roles.

Singularity container
A way to package up needed files into a single, verifiable file. This allows software developers to build their stack onto any operating system base.

SingularityNET
A framework for AGI research and development.

The Singularity of God
A term used by theologians to describe the idea that the god of the Bible is the one and only true God.

Some prominent technologists and academics dispute the plausibility of a technological singularity.